pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodocs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
 "google>=3.0.0",
 "requests>=2.32.4",
]

[tool.setuptools.packages.find]
where = ["main"]
include = ["*"]
namespaces = false

[tool.setuptools.package-dir]
"" = "main"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["main"]
extraPaths = ["main"]
